requirements.txt
pandas>=2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
tqdm>=4.66
psycopg2-binary
sentence-transformers
transformers
rapidfuzz
phonetics
requests
beautifulsoup4
lxml
nltk
